reviewer symantics and corrections
completed the following
You need to catch possible errors at the very end of any server request.
You need to toggle likes only in block then after a successful response from the server because if there is an error the user will not know about that and will see the toggled heart color.
There are 3 popups in the project with buttons and 2 forms. You need to name the elements the way all could understand what they are. And the functions as well
You need to close popups only in blocks then after a successful response from the server so that when there is an error it would be opened (the inputs will not be cleared and the user could send the data again to the server) and if you change the button text you could see the change as well
Class UserInfo should set the entire user’s data (including the avatar)